Adds wizard textures, fixed render bug
Found the best way to export materials was to seperate the UVMap parts
by material ID and then place colours manually in an image editor.

If there was a null texture loaded after all other textures, all
model textures would be null. This was fixed by checking if the texture
is null during the draw cycle and disabling the textures if it was.