Adds upgrades and persitant data
Upgrades are acquirable using gold. When acquired, an event will be
sent with a value. An object will need to be subscribed and have
a handler to make use of the value.

Data is saved to a persistantData.txt file in the format:
int gold,
int highscore,
intTier.intTier.intTier