Refactored minigames, added checklist and score
Stirring minigame will now appear when a recipe is ready to be made.
Grinding minigame will now appear when inserting an item.
Score will be generated based on the time taken to complete minigames.
Score is currently displayed in the top left corner of the screen.

Checklist is a class that takes in a recipe and will generate a list
of ingredients that need to marked as done. This class will help ensure
that the score cannot be exploited. Will also be used to deduct points
when items are put into the bin (reseting the score for that item).