Adds sliding minigame, bug fixes, UI, player lock
Added a minigame for the liquifying table with UI. Currently only
works for keyboard.

Added a lock event for the player in which the player will not be able
to move. Minigames send this event when started.

Fixed a bug where stirring minigame could only be played once.
Fixed a bug that caused the game to crash when trying to make the
2nd potion.