Refactored score system, processing, adds progress
Player can now convert a solid into either a dust or a liquid. Removed
the need to convert from sold to dust before the player could get
a liquid.

Score system was a bit of a mess with events being sent everywhere.
Made a sort of central hub for the score.

Player can now deposit potions in the chest and recieve the next recipe.